Writing flow: extract tab behaviour to hook
