--- packages/block-editor/src/components/writing-flow/use-last-focus.js
@@ -1,17 +1,0 @@
-/**
- * WordPress dependencies
- */
-import { useRefEffect } from '@wordpress/compose';
-
-export default function useLastFocus( lastFocus ) {
-	return useRefEffect( ( node ) => {
-		function onFocusOut( event ) {
-			lastFocus.current = event.target;
-		}
-
-		node.addEventListener( 'focusout', onFocusOut );
-		return () => {
-			node.removeEventListener( 'focusout', onFocusOut );
-		};
-	}, [] );
-}
